+++ HEN_HOUSE/egs++/egs_base_source.h
@@ -46,6 +46,7 @@
 #include "egs_vector.h"
 #include "egs_object_factory.h"
 #include "egs_functions.h"
+#include "egs_ensdf.h"
 
 #include <string>
 #include <iostream>
@@ -193,10 +194,16 @@ public:
 
     //virtual void setTimeIndex(EGS_Float temp_time) {};
 
-    /* A virtual function that prints statistics on the emissions from a source.
-    * Currently this is only implemented for EGS_RadionuclideSource.
-    */
+    /*!  \brief Print statistics on what was sampled from the source. */
     virtual void printSampledEmissions() {};
+
+   /*!  \brief Get the radionuclide ENSDF object from the source.
+    *
+    * This is only defined for EGS_RadionuclideSource
+    */
+    virtual vector<EGS_Ensdf*> getRadionuclideEnsdf() {
+        return vector<EGS_Ensdf*>();
+    };
 
     /*!  \brief Store the source state into the stream \a data_out.
      *
